refactor: OpenAi DTO KaKao DTO와 일관되게 리팩토링
